rptest: Break out partition movement utils
- Moving them into a new file

(cherry picked from commit 98c2a6736ec2b67362fc34a13ec8252eb32dc048)
