+++ tests/rptest/tests/partition_movement_test.py
@@ -20,13 +20,14 @@ from rptest.clients.types import TopicSpec
 from rptest.clients.rpk import RpkTool
 from rptest.tests.end_to_end import EndToEndTest
 from rptest.services.admin import Admin
+from rptest.tests.partition_movement import PartitionMovementMixin
 from rptest.services.honey_badger import HoneyBadger
 from rptest.services.rpk_producer import RpkProducer
 from rptest.services.kaf_producer import KafProducer
 from rptest.services.rpk_consumer import RpkConsumer
 
 
-class PartitionMovementTest(EndToEndTest):
+class PartitionMovementTest(PartitionMovementMixin, EndToEndTest):
     """
     Basic partition movement tests. Each test builds a number of topics and then
     performs a series of random replica set changes. After each change a
@@ -48,119 +49,6 @@ class PartitionMovementTest(EndToEndTest):
             },
             **kwargs)
         self._ctx = ctx
-
-    @staticmethod
-    def _random_partition(metadata):
-        topic = random.choice(metadata)
-        partition = random.choice(topic["partitions"])
-        return topic["topic"], partition["partition"]
-
-    @staticmethod
-    def _choose_replacement(admin, assignments):
-        """
-        Does not produce assignments that contain duplicate nodes. This is a
-        limitation in redpanda raft implementation.
-        """
-        replication_factor = len(assignments)
-        node_ids = lambda x: set([a["node_id"] for a in x])
-
-        assert replication_factor >= 1
-        assert len(node_ids(assignments)) == replication_factor
-
-        # remove random assignment(s). we allow no changes to be made to
-        # exercise the code paths responsible for dealing with no-ops.
-        num_replacements = random.randint(0, replication_factor)
-        selected = random.sample(assignments, num_replacements)
-        for assignment in selected:
-            assignments.remove(assignment)
-
-        # choose a valid random replacement
-        replacements = []
-        brokers = admin.get_brokers()
-        while len(assignments) != replication_factor:
-            broker = random.choice(brokers)
-            node_id = broker["node_id"]
-            if node_id in node_ids(assignments):
-                continue
-            core = random.randint(0, broker["num_cores"] - 1)
-            replacement = dict(node_id=node_id, core=core)
-            assignments.append(replacement)
-            replacements.append(replacement)
-
-        return selected, replacements
-
-    @staticmethod
-    def _get_assignments(admin, topic, partition):
-        res = admin.get_partitions(topic, partition)
-
-        def normalize(a):
-            return dict(node_id=a["node_id"], core=a["core"])
-
-        return [normalize(a) for a in res["replicas"]]
-
-    @staticmethod
-    def _equal_assignments(r0, r1):
-        def to_tuple(a):
-            return a["node_id"], a["core"]
-
-        r0 = [to_tuple(a) for a in r0]
-        r1 = [to_tuple(a) for a in r1]
-        return set(r0) == set(r1)
-
-    def _get_current_partitions(self, admin, topic, partition_id):
-        def keep(p):
-            return p["ns"] == "kafka" and p["topic"] == topic and p[
-                "partition_id"] == partition_id
-
-        result = []
-        for node in self.redpanda.nodes:
-            node_id = self.redpanda.idx(node)
-            partitions = admin.get_partitions(node=node)
-            partitions = filter(keep, partitions)
-            for partition in partitions:
-                result.append(dict(node_id=node_id, core=partition["core"]))
-        return result
-
-    def _move_and_verify(self):
-        admin = Admin(self.redpanda)
-
-        # choose a random topic-partition
-        metadata = self.redpanda.describe_topics()
-        topic, partition = self._random_partition(metadata)
-        self.logger.info(f"selected topic-partition: {topic}-{partition}")
-
-        # get the partition's replica set, including core assignments. the kafka
-        # api doesn't expose core information, so we use the redpanda admin api.
-        assignments = self._get_assignments(admin, topic, partition)
-        self.logger.info(f"assignments for {topic}-{partition}: {assignments}")
-
-        # build new replica set by replacing a random assignment
-        selected, replacements = self._choose_replacement(admin, assignments)
-        self.logger.info(
-            f"replacement for {topic}-{partition}:{len(selected)}: {selected} -> {replacements}"
-        )
-        self.logger.info(
-            f"new assignments for {topic}-{partition}: {assignments}")
-
-        r = admin.set_partition_replicas(topic, partition, assignments)
-
-        def status_done():
-            info = admin.get_partitions(topic, partition)
-            self.logger.info(
-                f"current assignments for {topic}-{partition}: {info}")
-            converged = self._equal_assignments(info["replicas"], assignments)
-            return converged and info["status"] == "done"
-
-        # wait until redpanda reports complete
-        wait_until(status_done, timeout_sec=90, backoff_sec=2)
-
-        def derived_done():
-            info = self._get_current_partitions(admin, topic, partition)
-            self.logger.info(
-                f"derived assignments for {topic}-{partition}: {info}")
-            return self._equal_assignments(info, assignments)
-
-        wait_until(derived_done, timeout_sec=90, backoff_sec=2)
 
     @cluster(num_nodes=3)
     def test_moving_not_fully_initialized_partition(self):
